Fixed the bugs

+++ technical-documents/source-code/leaderboard/models.py
@@ -9,6 +9,6 @@ class BuildingModel(models.Model):
     name = models.CharField(choices=building_choices, max_length=48)
     stats_since = models.DateField(
         default=datetime.today().strftime('%Y-%m-%d'))
-    co2 = models.PositiveIntegerField(default=0)
+    co2 = models.FloatField(default=0)
     number_submissions = models.IntegerField(default=0)
     last_done = models.DateTimeField(default='2023-01-01 00:00:00 +0000')
